jsonptr::assign

Type Alias AssignError

pub type AssignError = Error;

👎Deprecated since 0.7.0: renamed to Error

Alias for Error - indicates a value assignment failed.

Aliased Type

pub enum AssignError {
    FailedToParseIndex {
        position: usize,
        offset: usize,
        source: ParseIndexError,
    },
    OutOfBounds {
        position: usize,
        offset: usize,
        source: OutOfBoundsError,
    },
}

Variants

FailedToParseIndex

A Token within the Pointer failed to be parsed as an array index.

Fields

position: usize

Position (index) of the token which failed to parse as an Index

offset: usize

Offset of the partial pointer starting with the invalid index.

source: ParseIndexError

The source ParseIndexError

OutOfBounds

A Token within the Pointer contains an Index which is out of bounds.

The current or resulting array’s length is less than the index.

Fields

position: usize

Position (index) of the token which failed to parse as an Index

offset: usize

Offset of the partial pointer starting with the invalid index.

source: OutOfBoundsError

The source OutOfBoundsError