jsonptr::delete

Trait Delete

pub trait Delete {
    type Value;

    // Required method
    fn delete(&mut self, ptr: &Pointer) -> Option<Self::Value>;
}

Delete is implemented by types which can internally remove a value based on a JSON Pointer

Required Associated Types

type Value

The type of value that this implementation can operate on.

Required Methods

fn delete(&mut self, ptr: &Pointer) -> Option<Self::Value>

Attempts to internally delete a value based upon a Pointer.