jsonptr

Struct PointerBuf

pub struct PointerBuf(/* private fields */);

An owned, mutable Pointer (akin to String).

This type provides methods like PointerBuf::push_back and PointerBuf::replace that mutate the pointer in place. It also implements core::ops::Deref to Pointer, meaning that all methods on Pointer slices are available on PointerBuf values as well.

Implementations

impl PointerBuf

pub fn root() -> Self

Creates a new PointerBuf pointing to a document root.

This is an alias to Self::new.

pub fn new() -> Self

Creates a new PointerBuf pointing to a document root.

pub unsafe fn new_unchecked(s: impl Into<String>) -> Self

Create a PointerBuf from a string that is known to be correctly encoded.

Safety

The provided string must adhere to RFC 6901.

pub fn parse(s: impl Into<String>) -> Result<Self, RichParseError>

Attempts to parse a string into a PointerBuf.

Errors

Returns a RichParseError if the string is not a valid JSON Pointer.

pub fn from_tokens<'t>(tokens: impl IntoIterator<Item: Into<Token<'t>>>) -> Self

Creates a new PointerBuf from a slice of non-encoded strings.

pub fn as_ptr(&self) -> &Pointer

Coerces to a Pointer slice.

pub fn push_front<'t>(&mut self, token: impl Into<Token<'t>>)

Pushes a Token onto the front of this Pointer.

pub fn push_back<'t>(&mut self, token: impl Into<Token<'t>>)

Pushes a Token onto the back of this Pointer.

pub fn pop_back(&mut self) -> Option<Token<'static>>

Removes and returns the last Token in the Pointer if it exists.

pub fn pop_front(&mut self) -> Option<Token<'static>>

Removes and returns the first Token in the Pointer if it exists.

pub fn append<P: AsRef<Pointer>>(&mut self, other: P) -> &PointerBuf

Merges two Pointers by appending other onto self.

pub fn replace<'t>( &mut self, index: usize, token: impl Into<Token<'t>>, ) -> Result<Option<Token<'_>>, ReplaceError>

Attempts to replace a Token by the index, returning the replaced Token if it already exists. Returns None otherwise.

Errors

A [ReplaceError] is returned if the index is out of bounds.

pub fn clear(&mut self)

Clears the Pointer, setting it to root ("").