Trait jsonptr::Assign

pub trait Assign {
    type Error;

    // Required method
    fn assign<V>(
        &mut self,
        ptr: &Pointer,
        value: V,
    ) -> Result<Assignment<'_>, Self::Error>
       where V: Into<Value>;
}

Assign is implemented by types which can internally assign a serde_json::Value by a JSON Pointer.

Required Associated Types

type Error

Error associated with Assign

Required Methods

fn assign<V>( &mut self, ptr: &Pointer, value: V, ) -> Result<Assignment<'_>, Self::Error>
where V: Into<Value>,

Assign a value of based on the path provided by a JSON Pointer.

Object Safety

This trait is not object safe.