Trait jsonptr::Delete

pub trait Delete {
    type Error;

    // Required method
    fn delete(&mut self, ptr: &Pointer) -> Option<Value>;
}

Delete is implemented by types which can internally remove a value based on a JSON Pointer

Required Associated Types

type Error

Error associated with Delete

Required Methods

fn delete(&mut self, ptr: &Pointer) -> Option<Value>

Attempts to internally delete a value based upon a Pointer.