jsonptr::resolve

Trait Resolve

pub trait Resolve {
    type Value;
    type Error;

    // Required method
    fn resolve(&self, ptr: &Pointer) -> Result<&Self::Value, Self::Error>;
}

A trait implemented by types which can resolve a reference to a value type from a path represented by a JSON Pointer.

Required Associated Types

type Value

The type of value that this implementation can operate on.

type Error

Error associated with Resolve

Required Methods

fn resolve(&self, ptr: &Pointer) -> Result<&Self::Value, Self::Error>

Resolve a reference to Self::Value based on the path in a Pointer.

Errors

Returns a Self::Error if the Pointer can not be resolved.