jsonptr::resolve

Trait ResolveMut

pub trait ResolveMut {
    type Value;
    type Error;

    // Required method
    fn resolve_mut(
        &mut self,
        ptr: &Pointer,
    ) -> Result<&mut Self::Value, Self::Error>;
}

A trait implemented by types which can resolve a mutable reference to a value type from a path represented by a JSON Pointer.

Required Associated Types

type Value

The type of value that is being resolved.

type Error

Error associated with ResolveMut

Required Methods

fn resolve_mut( &mut self, ptr: &Pointer, ) -> Result<&mut Self::Value, Self::Error>

Resolve a mutable reference to a serde_json::Value based on the path in a JSON Pointer.

Errors

Returns a Self::Error if the Pointer can not be resolved.